rename sandbox files
