+++ src/components/icon-file/icon-file.sandbox.html
