+++ src/components/icon/icon.sandbox.ts
@@ -2,7 +2,7 @@ import Vue, { PluginObject } from 'vue';
 import { Component } from 'vue-property-decorator';
 
 import { ICON_NAME } from '../component-names';
-import WithRender from './icon-sandbox.html';
+import WithRender from './icon.sandbox.html';
 
 @WithRender
 @Component
