+++ src/components/sandbox/index.ts
@@ -1,7 +1,7 @@
 import Vue, { PluginObject } from 'vue';
 
-import IconFileSandboxPlugin from '../icon-file/icon-file-sandbox';
-import IconSandboxPlugin from '../icon/icon-sandbox';
+import IconFileSandboxPlugin from '../icon-file/icon-file.sandbox';
+import IconSandboxPlugin from '../icon/icon.sandbox';
 import RadioSandboxPlugin from '../radio/radio-sandbox';
 
 const SandboxesPlugin: PluginObject<any> = {
